Add CSV export support
Former-commit-id: 4fa78f272901ebb943fcc5be931bafd7d86b1c1c

+++ spatial/python/planet/planet_client.py
@@ -45,6 +45,7 @@ class PClientV1():
         self.maximgs = int(imagery_config['maximgs'])  # 15 #10 #20
         self.output_encoding = imagery_config['output_encoding']
         self.output_filename = imagery_config['output_filename']
+        self.output_filename_csv = imagery_config['output_filename_csv']
         self.catalog_path = imagery_config['catalog_path']
         self.s3_catalog_bucket = imagery_config['s3_catalog_bucket']
         self.s3_catalog_prefix = imagery_config['s3_catalog_prefix']
@@ -201,17 +202,33 @@ class PClientV1():
 
         return result_path
 
+    def upload_s3_csv(self):
+        output_key = "{}/{}".format(self.s3_catalog_prefix, self.output_filename)
+        result = 's3://{}/{}'.format(self.s3_catalog_bucket, output_key)
+        self.transfer.upload_file(self.output_filename, self.s3_catalog_bucket, output_key)
+        return result
+
+    def upload_s3_csv_csv(self):
+        output_key = "{}/{}".format(self.s3_catalog_prefix, self.output_filename_csv)
+        result = 's3://{}/{}'.format(self.s3_catalog_bucket, output_key)
+        self.transfer.upload_file(self.output_filename_csv, self.s3_catalog_bucket, output_key)
+        return result
+
     def download_localfs_s3(self, scene_id, season = ''):
         filepath = ''
-        output_key = "{}/{}/{}.tif".format(self.s3_catalog_prefix, season, scene_id)
-        result = 's3://{}/{}'.format(self.s3_catalog_bucket, output_key)
+        output_key = '{}/{}/{}.tif'.format(self.s3_catalog_prefix, season, scene_id)
+        s3_result = 's3://{}/{}'.format(self.s3_catalog_bucket, output_key)
+        local_result = '{}{}/{}.tif'.format(self.catalog_path, season, scene_id)
 
-        try:
-            self.s3client.head_object(Bucket = self.s3_catalog_bucket, Key = output_key)
-            filepath = result
-        except botocore.exceptions.ClientError:
-            filepath = self.download_localfs(scene_id, season)
-            self.logger.info("Uploading {}...".format(scene_id))
-            self.transfer.upload_file(filepath, self.s3_catalog_bucket, output_key)
+        if not os.path.exists(local_result):
+            try:
+                self.s3client.head_object(Bucket = self.s3_catalog_bucket, Key = output_key)
+                filepath = result
+            except botocore.exceptions.ClientError:
+                filepath = self.download_localfs(scene_id, season)
+                self.logger.info("Uploading {}...".format(scene_id))
+                self.transfer.upload_file(filepath, self.s3_catalog_bucket, output_key)
+        else:
+            filepath = local_result
 
-        return filepath, result
+        return filepath, s3_result
